revised the code to store in myComposable.kt
