+++ .idea/deploymentTargetSelector.xml
@@ -5,10 +5,13 @@
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
-      <SelectionState runConfigName="JoinEventPagePreview">
+      <SelectionState runConfigName="MapPagePreview">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
-      <SelectionState runConfigName="MapPagePreview">
+      <SelectionState runConfigName="PreviewLoginPage">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
+      <SelectionState runConfigName="CreateEventPageReview">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
     </selectionStates>
