+++ app/src/main/java/com/altaygunbatan/uni_verse/ui/ForgotPassword.kt
@@ -36,12 +36,14 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
+import com.altaygunbatan.uni_verse.R
 import com.altaygunbatan.uni_verse.ui.theme.bodyFontFamily
 import com.altaygunbatan.uni_verse.ui.theme.displayFontFamily
 
@@ -50,153 +52,27 @@ import com.google.firebase.auth.FirebaseAuth
 
 @Composable
 fun ForgotPasswordPage(navController : NavController) {
-    var email by rememberSaveable { mutableStateOf("") }
 
-    var showMessage by remember { mutableStateOf(false) }
 
-    var message by remember { mutableStateOf("") }
 
     Column(
         modifier = Modifier.fillMaxSize()
             .background(color = Color(red = 242, green = 244, blue = 243)),
     ) {
-        Text(
-            text = "FORGOT PASSWORD",
-            fontFamily = displayFontFamily,
-            fontSize = 32.sp,
-            modifier = Modifier.padding(
-                start = 30.dp,
-                top = 47.dp
-                )
-        )
 
-        Spacer(modifier = Modifier.height(30.dp))
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .wrapContentSize(Alignment.Center)
-        ) {
-            Text(
-                text = "No problem! Enter the e-mail address associated with your account. We'll send you a secure link to reset your password",
-                fontFamily = bodyFontFamily,
-                fontSize = 22.sp,
-                color = Color(red = 10, green = 16, blue = 69, alpha = 255),
-                modifier = Modifier.padding(start = 10.dp)
-            )
-        }
-        Spacer(modifier = Modifier.height(100.dp))
+        ForgotPasswordPageText()
 
-        Text(
-            text = "Register E-Mail",
-            fontFamily = bodyFontFamily,
-            fontSize = 22.sp,
-            modifier = Modifier.padding(start = 30.dp)
-        )
+        ForgotPasswordPageToSendLink()
 
-        Spacer(modifier = Modifier.height(35.dp))
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .wrapContentSize(Alignment.Center)
-        ) {
-            androidx.compose.material.TextField(
-                modifier = Modifier.size(width = 345.dp, height = 55.dp),
-                value = email,
-                onValueChange = {
-                    email = it
-                },
-                colors = TextFieldDefaults.textFieldColors(
-                    unfocusedIndicatorColor = Color.Transparent,
-                    focusedIndicatorColor = Color.Transparent
-                ),
-                shape = RoundedCornerShape(20.dp),
-
-                )
-
-        }
-        Spacer(modifier = Modifier.height(70.dp))
-
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .wrapContentSize(Alignment.Center)
-        ) {
-            Button(
-                onClick = {
-                    sendPasswordResetEmail(email) { success, response ->
-                        message = if (success) {
-                            "Reset link sent to $email"
-                        } else {
-                            response ?: "Error occurred, please try again."
-                        }
-                        showMessage = true
-                    }
-                },
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color(red = 10, green = 16, blue = 69, alpha = 255),
-
-                ),
-                shape = RoundedCornerShape(20.dp),
-                modifier = Modifier.size(width = 345.dp, height = 45.dp)
-            ) {
-                Text(
-                    text = "SEND LINK",
-                    color = Color.White,
-                    fontFamily = displayFontFamily,
-                    fontSize = 22.sp
-                )
-            }
-
-        }
-
-        Spacer(modifier = Modifier.height(180.dp))
-
-        TextButton(
-            onClick = {
-                navController.popBackStack()
-            },
-            modifier = Modifier.padding(start = 19.dp)
-        ) {
-            Text(
-                text = "Back",
-                fontFamily = displayFontFamily,
-                fontSize = 15.sp,
-                color = Color(red = 10, green = 16, blue = 69, alpha = 255)
-            )
-        }
+        ForgotPasswordPagrBackButton(navController = navController)
 
 
     }
-    if (showMessage) {
-        AlertDialog(
-            onDismissRequest = { showMessage = false },
-            title = { Text("Message") },
-            text = { Text(message) },
-            confirmButton = {
-                TextButton(onClick = { showMessage = false }) {
-                    Text("OK")
-                }
-            }
-        )
-    }
+
 }
 
 
-fun sendPasswordResetEmail(email: String, callback: (Boolean, String?) -> Unit) {
-    if (email.isEmpty()) {
-        callback(false, "Please enter your email.")
-        return
-    }
 
-    FirebaseAuth.getInstance().sendPasswordResetEmail(email)
-        .addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                callback(true, null)
-            } else {
-                callback(false, task.exception?.localizedMessage)
-            }
-        }
-}
 
 @Preview
 @Composable
