+++ app/src/main/java/com/altaygunbatan/uni_verse/ui/MyComposable.kt
@@ -1,5 +1,6 @@
 package com.altaygunbatan.uni_verse.ui
 
+import android.app.DatePickerDialog
 import android.net.Uri
 import android.os.Build
 import android.widget.Toast
@@ -13,7 +14,9 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
 
 
 import androidx.compose.foundation.layout.fillMaxSize
@@ -21,6 +24,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.foundation.lazy.LazyRow
 
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.items
@@ -30,10 +35,12 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.AlertDialog
+import androidx.compose.material.ButtonDefaults
 
 //noinspection UsingMaterialAndMaterial3Libraries
 import androidx.compose.material.Card
 import androidx.compose.material.Divider
+import androidx.compose.material.RadioButton
 
 //noinspection UsingMaterialAndMaterial3Libraries
 import androidx.compose.material.Text
@@ -64,6 +71,7 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -73,8 +81,14 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.text.style.TextAlign
 
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -86,55 +100,501 @@ import coil.compose.rememberAsyncImagePainter
 import coil.compose.rememberImagePainter
 import com.altaygunbatan.uni_verse.R
 import com.altaygunbatan.uni_verse.dataClasses.Event
+import com.altaygunbatan.uni_verse.ui.theme.bodyFontFamily
 import com.altaygunbatan.uni_verse.ui.theme.displayFontFamily
 import com.altaygunbatan.uni_verse.viewModels.EventViewModel
 
 import com.google.firebase.auth.FirebaseAuth
 import java.time.LocalDate
+import java.util.Calendar
 
 
 @Composable
-fun EmailTextField() {
+fun LoginPageEmailText(modifier: Modifier = Modifier) {
 
-        var email by remember { mutableStateOf("") }
-    TextField(
-        modifier = Modifier.size(width = 345.dp, height = 47.dp),
-        value =email,
-        onValueChange = {
-            email = it
-        },
-
+    Text(
+        text = stringResource(id = R.string.login),
+        fontFamily = displayFontFamily,
+        fontSize = 32.sp,
+        modifier = Modifier.padding(start = 32.dp, top = 45.dp,)
     )
+
+    Spacer(modifier = Modifier.height(120.dp))
+
+    Text(
+        text = stringResource(id = R.string.email),
+        fontFamily = bodyFontFamily,
+        fontSize = 22.sp,
+        color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+        modifier = Modifier.padding(start = 46.dp))
+
+    Spacer(modifier = Modifier.height(20.dp))
 }
 
 @Composable
-fun PasswordTextField() {
-    var password by remember { mutableStateOf("") }
+fun LoginToHomePage(modifier: Modifier = Modifier, onLoginSuccess: () -> Unit) {
+    var errorMessage by remember { mutableStateOf("") }
 
-    TextField(
-        modifier = Modifier.size(width = 345.dp, height = 45.dp),
-        value =password,
-        onValueChange = {
-            password = it
-        },
-        label = {
-            Text(text = "Password")
-        }
-    )
-}
+    var email by rememberSaveable { mutableStateOf("") }
+    var passwordVisibility by remember { mutableStateOf(false)}
 
-
-
-@Composable
-fun MyButton( text: String) {
-
-    Button(onClick = {},
-        modifier = Modifier
-            .size(width = 345.dp, height = 45.dp)
-        )
-    {
-        Text(text)
+    val icon = if (passwordVisibility){
+        painterResource(id = R.drawable.baseline_visibility_24)
     }
+    else {
+        painterResource(id = R.drawable.baseline_visibility_off_24)
+    }
+    var password by rememberSaveable { mutableStateOf("") }
+
+    var isSelected by rememberSaveable { mutableStateOf(false) }
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ){
+        androidx.compose.material.TextField(
+            modifier = Modifier.size(width = 345.dp, height = 55.dp),
+            value = email,
+            onValueChange = {
+                email = it
+            },
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(20.dp)
+        )
+    }
+
+    Spacer(modifier = Modifier.height(50.dp))
+
+
+    Text(
+        text = stringResource(id = R.string.password),
+        fontFamily = bodyFontFamily,
+        fontSize = 22.sp,
+        color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+        modifier = Modifier.padding(start = 46.dp)
+    )
+
+    Spacer(modifier = Modifier.height(20.dp))
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        androidx.compose.material.TextField(
+            modifier = Modifier.size(width = 345.dp, height = 55.dp),
+            value = password,
+            onValueChange = {
+                password = it
+            },
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(20.dp),
+            trailingIcon = {
+                IconButton(onClick = {
+                    passwordVisibility = !passwordVisibility
+                }) {
+                    Icon(
+                        painter = icon,
+                        contentDescription = "Click to change password visibility"
+                    )
+                }
+            }, keyboardOptions = KeyboardOptions(
+                keyboardType = KeyboardType.Password
+            ),
+            visualTransformation = if (passwordVisibility) VisualTransformation.None
+            else PasswordVisualTransformation()
+        )
+    }
+    Spacer(modifier = Modifier.height(30.dp))
+
+    Row (verticalAlignment = Alignment.CenterVertically){
+        RadioButton(
+            selected = isSelected,
+            onClick = { isSelected = !isSelected },
+            modifier = Modifier.padding(start = 10.dp)
+        )
+        Text(
+            stringResource(id = R.string.remember),
+            fontFamily = bodyFontFamily,
+            color = Color(red = 10, green = 16, blue = 69, alpha = 255))
+    }
+
+    Spacer(modifier = Modifier.height(20.dp))
+
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+
+        androidx.compose.material.Button(
+            onClick = {
+                signInWithEmailAndPassword(email, password) { success, error ->
+                    if (success) {
+                        onLoginSuccess()
+                    } else {
+                        errorMessage = error ?: "An unknown error occurred"
+                    }
+                }
+            },
+            colors = androidx.compose.material.ButtonDefaults.buttonColors(
+                backgroundColor = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                contentColor = Color.White
+            ),
+            shape = RoundedCornerShape(20.dp),
+            modifier = Modifier
+                .size(width = 345.dp, height = 45.dp)
+        )
+        {
+            Text(
+                text = stringResource(R.string.signIn),
+                fontFamily = displayFontFamily,
+                fontSize = 22.sp
+            )
+        }
+    }
+
+    Spacer(modifier = Modifier.height(10.dp))
+
+}
+
+@Composable
+fun AccountForgotText(modifier: Modifier = Modifier, navController: NavController) {
+
+    Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.CenterEnd) {
+        TextButton(onClick = {
+            navController.navigate("forgot")
+        }
+        )
+        {
+            Text(text = stringResource(R.string.forgot),
+                fontFamily = displayFontFamily,
+                fontSize = 16.sp,
+                color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                modifier = Modifier.padding(end = 20.dp))
+
+        }
+    }
+    Spacer(modifier = Modifier.height(10.dp))
+    Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.CenterEnd) {
+        TextButton(onClick = {
+            navController.navigate("signup")
+        }) {
+            Text(
+                stringResource(R.string.account),
+                fontFamily = displayFontFamily,
+                fontSize = 16.sp,
+                color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                modifier = Modifier.padding(end = 20.dp))
+
+        }
+    }
+
+}
+
+@Composable
+fun SignUpPageEmailText(modifier: Modifier = Modifier) {
+
+    Text(
+        text = stringResource(id = R.string.signUp),
+        fontFamily = displayFontFamily,
+        color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+        fontSize = 32.sp,
+        modifier = Modifier.padding(start = 32.dp, top = 45.dp)
+    )
+
+    Spacer(modifier = Modifier.height(120.dp))
+
+    Text(
+        text = stringResource(id = R.string.email),
+        fontFamily = bodyFontFamily,
+        fontSize = 22.sp,
+        modifier = Modifier.padding(start = 46.dp))
+
+    Spacer(modifier = Modifier.height(20.dp))
+}
+
+@Composable
+fun SignUpPageToLogin(modifier: Modifier = Modifier, onSignupSuccess: () -> Unit) {
+
+    var password by rememberSaveable { mutableStateOf("") }
+    var email by rememberSaveable { mutableStateOf("") }
+    var errorMessage by remember { mutableStateOf("") }
+    var passwordVisibility by remember { mutableStateOf(false)}
+
+    val icon = if (passwordVisibility){
+        painterResource(id = R.drawable.baseline_visibility_24)
+    }
+    else {
+        painterResource(id = R.drawable.baseline_visibility_off_24)
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        androidx.compose.material.TextField(
+            modifier = Modifier.size(width = 345.dp, height = 55.dp),
+            value = email,
+            onValueChange = {
+                email = it
+            },
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(20.dp)
+        )
+    }
+    Spacer(modifier = Modifier.height(50.dp))
+
+    Text(
+        text = stringResource(id = R.string.password),
+        fontFamily = bodyFontFamily,
+        fontSize = 22.sp,
+        modifier = Modifier.padding(start = 46.dp),
+
+        )
+    Spacer(modifier = Modifier.height(20.dp))
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ){
+
+        androidx.compose.material.TextField(
+            modifier = Modifier.size(width = 345.dp, height = 55.dp),
+            value = password,
+            onValueChange = {
+                password = it
+            },
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(20.dp),
+            trailingIcon = {
+                IconButton(onClick = {
+                    passwordVisibility = !passwordVisibility
+                }) {
+                    Icon(
+                        painter = icon,
+                        contentDescription = "Click to change password visibility"
+                    )
+                }
+            }, keyboardOptions = KeyboardOptions(
+                keyboardType = KeyboardType.Password
+            ),
+            visualTransformation = if (passwordVisibility) VisualTransformation.None
+            else PasswordVisualTransformation()
+        )
+    }
+
+    Spacer(modifier = Modifier.height(60.dp))
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        androidx.compose.material.Button(
+            onClick = {
+                signUpWithEmailAndPassword(email, password) { success, error ->
+                    if (success) {
+                        onSignupSuccess()
+                    } else {
+                        errorMessage = error ?: "An unknown error occurred"
+                    }
+                }
+            },
+            colors = ButtonDefaults.buttonColors(
+                backgroundColor = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                contentColor = Color.White
+            ),
+            shape = RoundedCornerShape(20.dp),
+            modifier = Modifier.size(width = 345.dp, height = 47.dp)
+        ) {
+            Text(
+                text = stringResource(R.string.createAccount),
+                fontFamily = displayFontFamily,
+                fontSize = 22.sp
+            )
+        }
+    }
+    if (errorMessage.isNotEmpty()) {
+        Text(
+            text = errorMessage,
+            color = Color.Red,
+            modifier = Modifier.padding(top = 8.dp)
+        )
+    }
+
+
+    Spacer(modifier = Modifier.height(30.dp))
+}
+
+@Composable
+fun PreviewOrLogin(modifier: Modifier = Modifier, navController: NavController) {
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        MyButton2(stringResource(id = R.string.preview), navController)
+
+    }
+    Spacer(modifier = Modifier.height(100.dp))
+
+    TextButton(onClick = {
+        navController.navigate("login") },
+        modifier = Modifier.padding(start = 19.dp)) {
+        Text(text = stringResource(id = R.string.back),
+            fontFamily = displayFontFamily,
+            fontSize = 15.sp,
+            color = Color(red = 10, green = 16, blue = 69, alpha = 255))
+    }
+
+}
+
+@Composable
+fun ForgotPasswordPageText(modifier: Modifier = Modifier) {
+
+
+
+    Text(
+        text = stringResource(id = R.string.forgot_password),
+        fontFamily = displayFontFamily,
+        fontSize = 32.sp,
+        modifier = Modifier.padding(
+            start = 30.dp,
+            top = 47.dp
+        )
+    )
+
+    Spacer(modifier = Modifier.height(30.dp))
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        Text(
+            text = stringResource(R.string.forgot_text),
+            fontFamily = bodyFontFamily,
+            fontSize = 22.sp,
+            color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+            modifier = Modifier.padding(start = 10.dp)
+        )
+    }
+    Spacer(modifier = Modifier.height(100.dp))
+
+    Text(
+        text = stringResource(R.string.register),
+        fontFamily = bodyFontFamily,
+        fontSize = 22.sp,
+        modifier = Modifier.padding(start = 30.dp)
+    )
+
+    Spacer(modifier = Modifier.height(35.dp))
+
+}
+
+@Composable
+fun ForgotPasswordPageToSendLink(modifier: Modifier = Modifier) {
+
+    var email by rememberSaveable { mutableStateOf("") }
+
+    var showMessage by remember { mutableStateOf(false) }
+
+    var message by remember { mutableStateOf("") }
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        androidx.compose.material.TextField(
+            modifier = Modifier.size(width = 345.dp, height = 55.dp),
+            value = email,
+            onValueChange = {
+                email = it
+            },
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent
+            ),
+            shape = RoundedCornerShape(20.dp),
+
+            )
+
+    }
+    Spacer(modifier = Modifier.height(70.dp))
+
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        Button(
+            onClick = {
+                sendPasswordResetEmail(email) { success, response ->
+                    message = if (success) {
+                        "Reset link sent to $email"
+                    } else {
+                        response ?: "Error occurred, please try again."
+                    }
+                    showMessage = true
+                }
+            },
+            colors = androidx.compose.material3.ButtonDefaults.buttonColors(
+                containerColor = Color(red = 10, green = 16, blue = 69, alpha = 255),
+
+                ),
+            shape = RoundedCornerShape(20.dp),
+            modifier = Modifier.size(width = 345.dp, height = 45.dp)
+        ) {
+            Text(
+                text = stringResource(id = R.string.send_link),
+                color = Color.White,
+                fontFamily = displayFontFamily,
+                fontSize = 22.sp
+            )
+        }
+
+    }
+
+    Spacer(modifier = Modifier.height(180.dp))
+
+
+
+}
+
+
+@Composable
+fun ForgotPasswordPagrBackButton(modifier: Modifier = Modifier, navController: NavController) {
+
+    TextButton(
+        onClick = {
+            navController.popBackStack()
+        },
+        modifier = Modifier.padding(start = 19.dp)
+    ) {
+        Text(
+            text = stringResource(id = R.string.back),
+            fontFamily = displayFontFamily,
+            fontSize = 15.sp,
+            color = Color(red = 10, green = 16, blue = 69, alpha = 255)
+        )
+    }
+
 }
 
 @Composable
@@ -193,29 +653,6 @@ fun HomeTextField(viewModel: EventViewModel){
 
 }
 
-fun signUpWithEmailAndPassword(email: String, password: String, onResult: (Boolean, String?) -> Unit) {
-    val auth = FirebaseAuth.getInstance()
-    auth.createUserWithEmailAndPassword(email, password)
-        .addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                onResult(true, null)
-            } else {
-                onResult(false, task.exception?.message)
-            }
-        }
-}
-
-fun signInWithEmailAndPassword(email: String, password: String, onResult: (Boolean, String?) -> Unit) {
-    val auth = FirebaseAuth.getInstance()
-    auth.signInWithEmailAndPassword(email, password)
-        .addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                onResult(true, null)
-            } else {
-                onResult(false, task.exception?.message)
-            }
-        }
-}
 
 
 
@@ -263,6 +700,58 @@ fun MyTopAppBar(navController: NavController, selected: MutableState<Int>) {
     )
 
     }
+
+@Composable
+fun HomePageTextField(modifier: Modifier = Modifier, viewModel: EventViewModel) {
+
+    Spacer(modifier = Modifier.height(10.dp))
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        HomeTextField(viewModel)
+
+    }
+    Spacer(modifier = Modifier.height(20.dp))
+
+    androidx.compose.material3.Text(
+        text = stringResource(R.string.my_events),
+        fontFamily = displayFontFamily,
+        fontSize = 20.sp,
+        style = TextStyle(fontWeight = FontWeight.Bold),
+        modifier = Modifier.padding(start = 30.dp)
+    )
+
+}
+
+@Composable
+fun HomePageDisplayEvents(modifier: Modifier = Modifier, viewModel: EventViewModel) {
+
+    val events by viewModel.events.collectAsState(initial = emptyList())
+
+    if (events.isEmpty()) {
+        androidx.compose.material3.Text(
+            text = "No events available",
+            modifier = Modifier.fillMaxSize(),
+            textAlign = TextAlign.Center
+        )
+    } else {
+        LazyRow(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp),
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            items(events) { event ->
+                EventCard(event, onDelete = { viewModel.deleteEvent(event) })
+            }
+        }
+
+
+    }
+
+}
 
 @Composable
 fun MyBottomAppBar(navController: NavController, selected: MutableState<Int>
@@ -339,14 +828,15 @@ fun MyBottomAppBar(navController: NavController, selected: MutableState<Int>
                             .fillMaxWidth()
                     ) {
                         Text(
-                            text = "Please select a page to go",
+                            text = stringResource(id = R.string.select_page),
                             style = MaterialTheme.typography.headlineSmall,
                             modifier = Modifier.padding(bottom = 8.dp),
                             color = Color.White
                         )
                         Divider()
                         Text(
-                            text = "Create Event",
+                            text = stringResource(R.string.create_event_page),
+                            fontFamily = bodyFontFamily,
                             color = Color.White,
                             modifier = Modifier
                                 .fillMaxWidth()
@@ -358,7 +848,8 @@ fun MyBottomAppBar(navController: NavController, selected: MutableState<Int>
                             fontSize = 18.sp
                         )
                         Text(
-                            text = "Join Event",
+                            text = stringResource(R.string.joint_event_page),
+                            fontFamily = bodyFontFamily,
                             color = Color.White,
                             modifier = Modifier
                                 .fillMaxWidth()
@@ -411,39 +902,201 @@ fun MyBottomAppBar(navController: NavController, selected: MutableState<Int>
 
 
 @Composable
-fun ImageUploadCard() {
+fun CreateEventPageImageUpload(modifier: Modifier = Modifier) {
     var eventImage by remember { mutableStateOf<Uri?>(null) }
-    val context = LocalContext.current
-
-    // Launcher to pick an image
 
     val launcher = rememberLauncherForActivityResult(ActivityResultContracts.GetContent()) { uri ->
         eventImage = uri
     }
-
-    Card(
+    Box(
         modifier = Modifier
-            .size(150.dp)
-            .clickable {
-                // Trigger the image picker when the card is clicked
-                launcher.launch("image/*")
-            },
-        shape = RoundedCornerShape(16.dp),
-        elevation = 4.dp
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
     ) {
-        eventImage?.let {
-            Image(
-                painter = rememberAsyncImagePainter(model = it),
-                contentDescription = "Event Image",
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(150.dp)
-                    .padding(top = 16.dp),
-                contentScale = ContentScale.Crop
-            )
-        }
+        Card(
+            modifier = Modifier
+                .size(200.dp)
+                .fillMaxWidth()
+                .aspectRatio(1f)
+                .clickable {
+                    // Trigger the image picker when the card is clicked
+                    launcher.launch("image/*")
+                },
+            shape = RoundedCornerShape(16.dp),
+            elevation = 4.dp,
+        ) {
+            eventImage?.let {
+                Image(
+                    painter = rememberAsyncImagePainter(model = it),
+                    contentDescription = "Event Image",
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                    contentScale = ContentScale.Crop
+                )
+            }
         }
     }
+}
+
+@Composable
+fun CreateEvent(modifier: Modifier = Modifier, navController: NavController, viewModel: EventViewModel) {
+
+    var eventName by remember { mutableStateOf("") }
+    var eventDetails by remember { mutableStateOf("") }
+    var eventDate by remember { mutableStateOf("") }
+    val context = LocalContext.current
+
+    val calendar = Calendar.getInstance()
+    val datePickerDialog = DatePickerDialog(
+        context,
+        { _, year, month, dayOfMonth ->
+            eventDate = "$dayOfMonth/${month + 1}/$year" // Format the date
+        },
+        calendar.get(Calendar.YEAR),
+        calendar.get(Calendar.MONTH),
+        calendar.get(Calendar.DAY_OF_MONTH)
+    )
+
+    var eventImage by remember { mutableStateOf<Uri?>(null) }
+
+    val launcher = rememberLauncherForActivityResult(ActivityResultContracts.GetContent()) { uri ->
+        eventImage = uri
+    }
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        Card(
+            modifier = Modifier
+                .size(200.dp)
+                .fillMaxWidth()
+                .aspectRatio(1f)
+                .clickable {
+                    // Trigger the image picker when the card is clicked
+                    launcher.launch("image/*")
+                },
+            shape = RoundedCornerShape(16.dp),
+            elevation = 4.dp,
+        ) {
+            eventImage?.let {
+                Image(
+                    painter = rememberAsyncImagePainter(model = it),
+                    contentDescription = "Event Image",
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(150.dp),
+                    contentScale = ContentScale.Crop
+                )
+            }
+        }
+    }
+
+
+    androidx.compose.material.TextField(
+        modifier = Modifier
+            .padding(16.dp),
+        colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedIndicatorColor = Color.Transparent,
+        ),
+        shape = RoundedCornerShape(20.dp),
+        value = eventName,
+        onValueChange = { eventName = it },
+        placeholder = {
+            androidx.compose.material3.Text(
+                text = stringResource(id = R.string.event_name),
+                color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                fontWeight = FontWeight.SemiBold
+            )
+        },
+    )
+    Row(
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        androidx.compose.material.TextField(
+            modifier = Modifier
+                .padding(16.dp),
+            colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+                unfocusedIndicatorColor = Color.Transparent,
+                focusedIndicatorColor = Color.Transparent,
+
+                ),
+            shape = RoundedCornerShape(20.dp),
+            value = eventDate,
+            onValueChange = { eventDate = it },
+            placeholder = {
+                androidx.compose.material3.Text(
+                    text = stringResource(id = R.string.event_date),
+                    color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                    fontWeight = FontWeight.SemiBold
+                )
+            },
+            enabled = false
+        )
+        IconButton(onClick = { datePickerDialog.show() }) {
+            Icon(
+                modifier = Modifier.size(30.dp),
+                painter = painterResource(id = R.drawable.fe_calendar),
+                contentDescription = "Select Date"
+            )
+        }
+    }
+
+
+    androidx.compose.material.TextField(
+        modifier = Modifier
+            .padding(16.dp),
+        colors = androidx.compose.material.TextFieldDefaults.textFieldColors(
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedIndicatorColor = Color.Transparent,
+
+            ),
+        shape = RoundedCornerShape(20.dp),
+        value = eventDetails,
+        onValueChange = { eventDetails = it },
+        placeholder = {
+            androidx.compose.material3.Text(
+                text = stringResource(id = R.string.event_details),
+                color = Color(red = 10, green = 16, blue = 69, alpha = 255),
+                fontWeight = FontWeight.SemiBold
+            )
+        },
+    )
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .wrapContentSize(Alignment.Center)
+    ) {
+        Button(
+            onClick = {
+                val event = Event(
+                    eventName = eventName,
+                    eventDetails = eventDetails,
+                    eventDate = eventDate,
+                    eventImage = eventImage?.toString(), // Save URI as a string
+
+                )
+                viewModel.addEvent(event)
+                navController.navigate("home")
+
+            },
+            colors = androidx.compose.material3.ButtonDefaults.buttonColors(
+                containerColor = Color(red = 255, green = 60, blue = 49, alpha = 255)
+            )
+        ) {
+            androidx.compose.material3.Text(
+                text = "Create",
+                fontFamily = displayFontFamily,
+                fontWeight = FontWeight.Bold,
+                fontSize = 25.sp
+            )
+        }
+    }
+
+}
 
 
 @Composable
