+++ app/src/main/java/com/altaygunbatan/uni_verse/ui/MyFunctions.kt
@@ -1,0 +1,45 @@
+package com.altaygunbatan.uni_verse.ui
+
+import com.google.firebase.auth.FirebaseAuth
+
+
+fun sendPasswordResetEmail(email: String, callback: (Boolean, String?) -> Unit) {
+    if (email.isEmpty()) {
+        callback(false, "Please enter your email.")
+        return
+    }
+
+    FirebaseAuth.getInstance().sendPasswordResetEmail(email)
+        .addOnCompleteListener { task ->
+            if (task.isSuccessful) {
+                callback(true, null)
+            } else {
+                callback(false, task.exception?.localizedMessage)
+            }
+        }
+}
+
+
+fun signUpWithEmailAndPassword(email: String, password: String, onResult: (Boolean, String?) -> Unit) {
+    val auth = FirebaseAuth.getInstance()
+    auth.createUserWithEmailAndPassword(email, password)
+        .addOnCompleteListener { task ->
+            if (task.isSuccessful) {
+                onResult(true, null)
+            } else {
+                onResult(false, task.exception?.message)
+            }
+        }
+}
+
+fun signInWithEmailAndPassword(email: String, password: String, onResult: (Boolean, String?) -> Unit) {
+    val auth = FirebaseAuth.getInstance()
+    auth.signInWithEmailAndPassword(email, password)
+        .addOnCompleteListener { task ->
+            if (task.isSuccessful) {
+                onResult(true, null)
+            } else {
+                onResult(false, task.exception?.message)
+            }
+        }
+}
